exclude sound controls from savestate
This fixes a sound queue overflow crash that occurs when a state is created on the same frame that a sound effect is put in $53. When loading that state, the APU will be effectively desynced from the CPU since we don't save the APU state, and the sound will never be processed.