prevent camera jerk on savestate load
Occasionally when loading a state, the column at the edge of the screen will not load. I think this is due to an incorrect camera position being set in gamemode 0C, and then suddenly jerking to the correct one that we load from the state. Gamemode 0C calculates a camera X based on Yoshi's X position, which we were restoring early for some reason.