fix uninitialised SRAM crashes
First, the loaded_state var could potentially be non-zero on boot, which would cause some bugs, so I moved it to an initialised SRAM region instead of a persistent one. Secondly, the debug_control_scheme persistent var was being used inside the main game loop hijack, and it could be basically any value instead of the valid 00 or 02. It is now moved to only run in game mode 0F and also is checked at boot, along with other persistent vars.