optimise custom bindings check
Don't enter the bindings loop unless a press (single frame) input is detected, since these are required for all custom binds. In terms of CPU cycles, we are in the "no input" state for the majority of the time:

check_bindings (no input): 32
check_bindings (pad1 input): 299 (each loop ~23 cycles) | 249 (only for 1st frame of any input)

The toggle music binding was removed since it adds a loop and it didn't seem useful.